savegame: split the module by what it owns
The savegame module was one file holding four separate concerns, and the
resume info being allocated next to the slot tables is what let a change
of slot count take the playthrough down with it.

Each part now sits in its own file:

- resume.c owns the resume info and is set up and taken down on its own
- manager.c owns the slot tables, the counters and the slot references
- common.c drives a save or a load of the running game
- vars.c holds the tables naming what a savegame carries

The headers follow, all of them reachable through the savegame.h
umbrella, so nothing outside the module changes. Savegame_Save keeps the
game-facing half beside Savegame_Load and asks the manager to write the
file.
