+++ src/meson.build
@@ -810,6 +810,9 @@ common_sources = [
   'trx/game/savegame/file.c',
   'trx/game/savegame/file_read.c',
   'trx/game/savegame/file_write.c',
+  'trx/game/savegame/manager.c',
+  'trx/game/savegame/resume.c',
+  'trx/game/savegame/vars.c',
   'trx/game/screenshot.c',
   'trx/game/shell/args.c',
   'trx/game/shell/common.c',
