+++ src/trx/game/savegame.h
@@ -2,4 +2,7 @@
 
 #include <trx/game/savegame/common.h>
 #include <trx/game/savegame/enum.h>
+#include <trx/game/savegame/manager.h>
+#include <trx/game/savegame/resume.h>
 #include <trx/game/savegame/types.h>
+#include <trx/game/savegame/vars.h>
