+++ src/trx/game/savegame/common.c
@@ -1,323 +1,16 @@
 #include <trx/config.h>
-#include <trx/core/benchmark.h>
-#include <trx/core/enum_map.h>
-#include <trx/core/memory.h>
-#include <trx/core/strings.h>
 #include <trx/debug.h>
-#include <trx/game/creature.h>
-#include <trx/game/cutseq.h>
 #include <trx/game/game.h>
 #include <trx/game/game_flow.h>
-#include <trx/game/gun.h>
-#include <trx/game/inventory.h>
+#include <trx/game/items.h>
 #include <trx/game/lara.h>
 #include <trx/game/objects.h>
-#include <trx/game/objects/traps/movable_block.h>
-#include <trx/game/pathing/lot.h>
-#include <trx/game/rules.h>
+#include <trx/game/rooms.h>
 #include <trx/game/savegame.h>
 #include <trx/game/savegame/file.h>
-#include <trx/game/shell.h>
 #include <trx/version.h>
 
-#include <stdio.h>
-#include <string.h>
-
-// Inventory items persisted in savegames, keyed by their legacy JSON names.
-const SAVEGAME_INVENTORY_ENTRY g_Savegame_InventoryItems[] = {
-// clang-format off
-#define X_PICKUP_KEY(n) { O_KEY_ITEM_##n, "key" #n },
-#define X_PICKUP_PUZZLE(n) { O_PUZZLE_ITEM_##n, "puzzle" #n },
-#define X_PICKUP_PICKUP(n) { O_PICKUP_ITEM_##n, "pickup" #n },
-#define X_PICKUP_QUEST(n) { O_QUEST_ITEM_##n, "quest" #n },
-#define X_PICKUP_KEY_COMBO(n, c)                                               \
-    { O_KEY_ITEM_##n##_COMBO_##c, "key" #n "_combo" #c },
-#define X_PICKUP_PUZZLE_COMBO(n, c)                                            \
-    { O_PUZZLE_ITEM_##n##_COMBO_##c, "puzzle" #n "_combo" #c },
-#define X_PICKUP_PICKUP_COMBO(n, c)                                            \
-    { O_PICKUP_ITEM_##n##_COMBO_##c, "pickup" #n "_combo" #c },
-#define X_PICKUP_EXAMINE(n) { O_EXAMINE_ITEM_##n, "examine" #n },
-#include <trx/game/objects/pickups.def>
-#undef X_PICKUP_EXAMINE
-#undef X_PICKUP_PICKUP_COMBO
-#undef X_PICKUP_PUZZLE_COMBO
-#undef X_PICKUP_KEY_COMBO
-#undef X_PICKUP_QUEST
-#undef X_PICKUP_PICKUP
-#undef X_PICKUP_PUZZLE
-#undef X_PICKUP_KEY
-    { O_LEADBAR_ITEM, "leadbar" },
-    { O_LASERSIGHT_ITEM, "lasersight" },
-    { O_BINOCULARS_ITEM, "binoculars" },
-    { O_CROWBAR_ITEM, "crowbar" },
-    { O_WATERSKIN_1_EMPTY, "waterskin1" },
-    { O_WATERSKIN_2_EMPTY, "waterskin2" },
-    { O_SAVE_CRYSTAL_ITEM, "save_crystal" },
-    { NO_OBJECT, nullptr },
-    // clang-format on
-};
-
-// Lara's ammunition as savegames name it.
-const SAVEGAME_AMMO_ENTRY g_Savegame_WeaponAmmo[] = {
-    // clang-format off
-    { LGT_PISTOLS,      "pistols",      true  },
-    { LGT_MAGNUMS,      "magnums",      true  },
-    { LGT_UZIS,         "uzis",         true  },
-    { LGT_SHOTGUN,      "shotgun",      true  },
-    { LGT_HARPOON,      "harpoon",      true  },
-    { LGT_GRENADE,      "grenade",      true  },
-    { LGT_M16,          "m16",          true  },
-    { LGT_AUTOS,        "autos",        false },
-    { LGT_DESERT_EAGLE, "desert_eagle", false },
-    { LGT_MP5,          "mp5",          false },
-    { LGT_ROCKET,       "rocket",       false },
-    { LGT_CROSSBOW,     "crossbow",     false },
-    { LGT_REVOLVER,     "revolver",     false },
-    { LGT_UNARMED,      nullptr,        false },
-    // clang-format on
-};
-
-// What a level keeps for Lara's return, under the names savegames give them.
-const SAVEGAME_RESUME_WEAPON g_Savegame_ResumeWeapons[] = {
-    // clang-format off
-    { LGT_PISTOLS,      "has_pistols",      "pistol_ammo",       true  },
-    { LGT_SHOTGUN,      "has_shotgun",      "shotgun_ammo",      true  },
-    { LGT_UZIS,         "has_uzis",         "uzi_ammo",          true  },
-    { LGT_M16,          "has_m16",          "m16_ammo",          true  },
-    { LGT_GRENADE,      "has_grenade",      "grenade_ammo",      true  },
-    { LGT_HARPOON,      "has_harpoon",      "harpoon_ammo",      true  },
-    { LGT_MAGNUMS,      "has_magnums",      "magnum_ammo",       true  },
-    // Introduced in TRX 1.1
-    { LGT_AUTOS,        "has_autos",        "autos_ammo",        false },
-    { LGT_DESERT_EAGLE, "has_desert_eagle", "desert_eagle_ammo", false },
-    { LGT_MP5,          "has_mp5",          "mp5_ammo",          false },
-    { LGT_ROCKET,       "has_rocket",       "rocket_ammo",       false },
-    // Introduced in TRX 1.9
-    { LGT_CROSSBOW,     "has_crossbow",     "crossbow_ammo",     false },
-    { LGT_REVOLVER,     "has_revolver",     "revolver_ammo",     false },
-    { LGT_UNARMED,      nullptr,            nullptr,             false },
-    // clang-format on
-};
-
-const SAVEGAME_RESUME_ITEM g_Savegame_ResumeItems[] = {
-    // clang-format off
-    { O_SMALL_MEDIPACK_ITEM, "num_medis",         true  },
-    { O_LARGE_MEDIPACK_ITEM, "num_big_medis",     true  },
-    { O_FLARE_ITEM,          "num_flares",        true  },
-    { O_SCION_ITEM_1,        "num_scions",        true  },
-    // Introduced in TRX 1.2
-    { O_QUEST_ITEM_1,        "num_quest_item_1",  false },
-    { O_QUEST_ITEM_2,        "num_quest_item_2",  false },
-    { O_QUEST_ITEM_3,        "num_quest_item_3",  false },
-    { O_QUEST_ITEM_4,        "num_quest_item_4",  false },
-    { O_QUEST_ITEM_5,        "num_quest_item_5",  false },
-    { O_QUEST_ITEM_6,        "num_quest_item_6",  false },
-    // Introduced in TRX 1.10
-    { O_SAVE_CRYSTAL_ITEM,   "num_save_crystals", false },
-    { NO_OBJECT,             nullptr,             false },
-    // clang-format on
-};
-
 static SAVEGAME_VERSION m_InitialVersion = SG_VERSION_LEGACY;
-static SAVEGAME_INFO *m_NormalSavegameInfo = nullptr;
-static SAVEGAME_INFO *m_QuickSavegameInfo = nullptr;
-static RESUME_INFO *m_ResumeInfo = nullptr;
-static int32_t m_SaveSlots = 0;
-static int32_t m_QuickSaveSlots = 0;
-static int32_t m_SavedGames = 0;
-static int32_t m_SaveCounter = 0;
-static int32_t m_NextQuickSlot = 0;
-static SAVEGAME_SLOT_REF m_MostRecentlyUsedSlot = { .index = -1 };
-static SAVEGAME_SLOT_REF m_MostRecentlyCreatedSlot = { .index = -1 };
-static SAVEGAME_SLOT_REF m_BoundSlot = { .index = -1 };
-
-static const char *M_GetSaveWriteDir(void)
-{
-    const char *const saves_dir = TRXPath_Get(TRX_PATH_SAVES_DIR);
-    const SHELL_ARGS *const args = Shell_GetArgs();
-    if (args != nullptr && args->startup.mod != nullptr
-        && args->startup.mod->name != nullptr) {
-        return String_FormatStatic("%s/%s", saves_dir, args->startup.mod->name);
-    }
-    return saves_dir;
-}
-
-static char *M_GetSaveWritePath(const char *const file_name)
-{
-    ASSERT(file_name != nullptr);
-    return String_Format("%s/%s", M_GetSaveWriteDir(), file_name);
-}
-
-static SAVEGAME_INFO *M_GetSavegameInfoSlot(const SAVEGAME_SLOT_REF slot)
-{
-    switch (slot.pool) {
-    case SAVEGAME_SLOT_POOL_NORMAL:
-        if (slot.index >= 0 && slot.index < m_SaveSlots) {
-            return &m_NormalSavegameInfo[slot.index];
-        }
-        break;
-    case SAVEGAME_SLOT_POOL_QUICK:
-        if (slot.index >= 0 && slot.index < m_QuickSaveSlots) {
-            return &m_QuickSavegameInfo[slot.index];
-        }
-        break;
-    case SAVEGAME_SLOT_POOL_NUMBER_OF:
-        break;
-    }
-    return nullptr;
-}
-
-static const char *M_GetSaveFilePatternForPool(const SAVEGAME_SLOT_POOL pool)
-{
-    switch (pool) {
-    case SAVEGAME_SLOT_POOL_NORMAL:
-        return SG_File_GetSaveFilePattern();
-    case SAVEGAME_SLOT_POOL_QUICK:
-        return SG_File_GetQuickSaveFilePattern();
-    case SAVEGAME_SLOT_POOL_NUMBER_OF:
-        break;
-    }
-    return nullptr;
-}
-
-static void M_CopyResumeInfo(
-    RESUME_INFO *const target, const RESUME_INFO *const source)
-{
-    memcpy(target, source, sizeof(RESUME_INFO));
-}
-
-// What a level keeps for Lara's return is what the savegame names: her
-// weapons and their ammunition, her supplies and her plot items. A key, a
-// puzzle piece or a waterskin belongs to the level she found it in, and is
-// left behind at its end.
-static void M_PersistInventory(RESUME_INFO *const resume)
-{
-    const INVENTORY_STATE *const live = Inv_GetState();
-    resume->inv = (INVENTORY_STATE) {};
-    memcpy(resume->inv.ammo, live->ammo, sizeof(resume->inv.ammo));
-
-    for (const SAVEGAME_RESUME_WEAPON *entry = g_Savegame_ResumeWeapons;
-         entry->has_key != nullptr; entry++) {
-        const OBJECT_ID gun_object = Gun_GetGunObject(entry->gun_type);
-        Inv_State_SetCount(
-            &resume->inv, gun_object, Inv_HasItem(gun_object) ? 1 : 0);
-    }
-    for (const SAVEGAME_RESUME_ITEM *entry = g_Savegame_ResumeItems;
-         entry->key != nullptr; entry++) {
-        Inv_State_SetCount(
-            &resume->inv, entry->object_id, Inv_GetItemCount(entry->object_id));
-    }
-    Inv_State_SetCount(
-        &resume->inv, O_BINOCULARS_ITEM,
-        Inv_HasItem(O_BINOCULARS_ITEM) ? 1 : 0);
-}
-
-static void M_ClearSlot(SAVEGAME_INFO *const savegame_info)
-{
-    savegame_info->counter = -1;
-    savegame_info->level_num = -1;
-    savegame_info->is_quick = false;
-    Memory_FreePointer(&savegame_info->full_path);
-    Memory_FreePointer(&savegame_info->level_title);
-}
-
-static void M_ClearSlots(void)
-{
-    if (m_NormalSavegameInfo != nullptr) {
-        for (int32_t i = 0; i < m_SaveSlots; i++) {
-            M_ClearSlot(&m_NormalSavegameInfo[i]);
-        }
-    }
-    if (m_QuickSavegameInfo != nullptr) {
-        for (int32_t i = 0; i < m_QuickSaveSlots; i++) {
-            M_ClearSlot(&m_QuickSavegameInfo[i]);
-        }
-    }
-}
-
-static void M_AllocSlots(void)
-{
-    m_SaveSlots = Savegame_GetSlotCount(SAVEGAME_SLOT_POOL_NORMAL);
-    m_QuickSaveSlots = Savegame_GetSlotCount(SAVEGAME_SLOT_POOL_QUICK);
-    m_NormalSavegameInfo = Memory_Alloc(sizeof(SAVEGAME_INFO) * m_SaveSlots);
-    m_QuickSavegameInfo = m_QuickSaveSlots > 0
-        ? Memory_Alloc(sizeof(SAVEGAME_INFO) * m_QuickSaveSlots)
-        : nullptr;
-}
-
-static void M_FreeSlots(void)
-{
-    M_ClearSlots();
-    Memory_FreePointer(&m_NormalSavegameInfo);
-    Memory_FreePointer(&m_QuickSavegameInfo);
-    m_SaveSlots = 0;
-    m_QuickSaveSlots = 0;
-}
-
-static bool M_FillSlot(const SAVEGAME_SLOT_REF slot, const char *const path)
-{
-    SAVEGAME_INFO *const savegame_info = M_GetSavegameInfoSlot(slot);
-    if (savegame_info == nullptr) {
-        return false;
-    }
-    bool result = false;
-    MYFILE *const fp = File_Open(path, FILE_OPEN_READ);
-    if (fp != nullptr) {
-        SAVEGAME_INFO tmp_savegame_info;
-        if (SG_File_FillInfo(fp, &tmp_savegame_info)) {
-            M_ClearSlot(savegame_info);
-            *savegame_info = tmp_savegame_info;
-            savegame_info->is_quick = slot.pool == SAVEGAME_SLOT_POOL_QUICK;
-            savegame_info->full_path = Memory_DupStr(path);
-            result = true;
-        }
-        File_Close(fp);
-    }
-    return result;
-}
-
-static void M_ScanSavedGamesDir(const char *const dir_path)
-{
-    void *const dir_handle = File_OpenDirectory(dir_path);
-    if (dir_handle == nullptr) {
-        return;
-    }
-
-    while (true) {
-        const char *const file_name = File_ReadDirectory(dir_handle);
-        if (file_name == nullptr) {
-            break;
-        }
-        if (strcmp(file_name, ".") == 0 || strcmp(file_name, "..") == 0) {
-            continue;
-        }
-
-        char *file_name_ci = String_ToUpper(file_name);
-        for (SAVEGAME_SLOT_POOL pool = 0; pool < SAVEGAME_SLOT_POOL_NUMBER_OF;
-             pool++) {
-            const char *const pattern = M_GetSaveFilePatternForPool(pool);
-            char *pattern_ci = String_ToUpperPattern(pattern);
-            int32_t slot_idx = -1;
-            const int32_t parsed = sscanf(file_name_ci, pattern_ci, &slot_idx);
-            Memory_FreePointer(&pattern_ci);
-
-            if (parsed != 1 || slot_idx < 0
-                || slot_idx >= Savegame_GetSlotCount(pool)) {
-                continue;
-            }
-
-            char *file_path = String_Format("%s/%s", dir_path, file_name);
-            M_FillSlot(
-                (SAVEGAME_SLOT_REF) { .pool = pool, .index = slot_idx },
-                file_path);
-            Memory_FreePointer(&file_path);
-            break;
-        }
-        Memory_FreePointer(&file_name_ci);
-    }
-
-    File_CloseDirectory(dir_handle);
-}
 
 static void M_LoadPreprocess(void)
 {
@@ -355,55 +48,6 @@ static void M_LoadPostprocess(void)
     }
 }
 
-static void M_DetermineLegacyGunTypes(RESUME_INFO *const resume)
-{
-    // Fallback logic to figure out holster and back gun items for saves from
-    // TR1X 4.2 and earlier (including TombATI) and TR2X 1.2 and earlier, where
-    // these values are missing. Make educated guesses based on the type of gun
-    // equipped.
-    if (resume->holsters_gun_type == LGT_UNKNOWN) {
-        switch (resume->equipped_gun_type) {
-        case LGT_PISTOLS:
-        case LGT_MAGNUMS:
-        case LGT_AUTOS:
-        case LGT_DESERT_EAGLE:
-        case LGT_UZIS:
-        case LGT_REVOLVER:
-            resume->holsters_gun_type = resume->equipped_gun_type;
-            break;
-        case LGT_SHOTGUN:
-        case LGT_M16:
-        case LGT_MP5:
-        case LGT_GRENADE:
-        case LGT_ROCKET:
-        case LGT_HARPOON:
-        case LGT_CROSSBOW:
-            resume->holsters_gun_type = Gun_GetHolsterChoice(&resume->inv);
-            break;
-        default:
-            resume->holsters_gun_type = LGT_UNARMED;
-            break;
-        }
-    }
-    if (resume->back_gun_type == LGT_UNKNOWN) {
-        resume->back_gun_type = Gun_GetBackChoice(&resume->inv);
-    }
-}
-
-static bool M_IsQuickSlotSortedBefore(
-    const SAVEGAME_SLOT_REF left, const SAVEGAME_SLOT_REF right)
-{
-    const SAVEGAME_INFO *const left_info = Savegame_GetSavegameInfo(left);
-    const SAVEGAME_INFO *const right_info = Savegame_GetSavegameInfo(right);
-    if (left_info == nullptr || right_info == nullptr) {
-        return false;
-    }
-    if (left_info->counter != right_info->counter) {
-        return left_info->counter > right_info->counter;
-    }
-    return left.index < right.index;
-}
-
 SAVEGAME_VERSION Savegame_GetInitialVersion(void)
 {
     return m_InitialVersion;
@@ -414,469 +58,16 @@ void Savegame_SetInitialVersion(const SAVEGAME_VERSION version)
     m_InitialVersion = version;
 }
 
-void Savegame_BindSlot(const SAVEGAME_SLOT_REF slot)
-{
-    if (!Savegame_IsValidSlotRef(slot)) {
-        m_BoundSlot = Savegame_InvalidSlot();
-        return;
-    }
-    m_BoundSlot = slot;
-    m_MostRecentlyUsedSlot = slot;
-    LOG_DEBUG("Binding save slot %d:%d", slot.pool, slot.index);
-}
-
-SAVEGAME_SLOT_REF Savegame_GetMostRecentlyUsedSlot(void)
-{
-    return m_MostRecentlyUsedSlot;
-}
-
-void Savegame_UnbindSlot(void)
-{
-    LOG_DEBUG("Resetting the save slot");
-    m_BoundSlot = Savegame_InvalidSlot();
-}
-
-SAVEGAME_SLOT_REF Savegame_GetBoundSlot(void)
-{
-    return m_BoundSlot;
-}
-
-int32_t Savegame_GetLevelNumber(const SAVEGAME_SLOT_REF slot)
-{
-    const SAVEGAME_INFO *const info = Savegame_GetSavegameInfo(slot);
-    return info != nullptr ? info->level_num : -1;
-}
-
-bool Savegame_IsSlotFree(const SAVEGAME_SLOT_REF slot)
-{
-    const SAVEGAME_INFO *const info = Savegame_GetSavegameInfo(slot);
-    return info == nullptr || info->level_num == -1;
-}
-
-int32_t Savegame_GetCounter(void)
-{
-    return m_SaveCounter;
-}
-
-int32_t Savegame_GetTotalCount(void)
-{
-    return m_SavedGames;
-}
-
-SAVEGAME_SLOT_REF Savegame_GetMostRecentlyCreatedSlot(void)
-{
-    return m_MostRecentlyCreatedSlot;
-}
-
-SAVEGAME_SLOT_REF Savegame_NormalSlot(const int32_t index)
-{
-    return (SAVEGAME_SLOT_REF) {
-        .pool = SAVEGAME_SLOT_POOL_NORMAL,
-        .index = index,
-    };
-}
-
-SAVEGAME_SLOT_REF Savegame_QuickSlot(const int32_t index)
-{
-    return (SAVEGAME_SLOT_REF) {
-        .pool = SAVEGAME_SLOT_POOL_QUICK,
-        .index = index,
-    };
-}
-
-SAVEGAME_SLOT_REF Savegame_InvalidSlot(void)
-{
-    return (SAVEGAME_SLOT_REF) {
-        .pool = SAVEGAME_SLOT_POOL_NORMAL,
-        .index = -1,
-    };
-}
-
-bool Savegame_IsValidSlotRef(const SAVEGAME_SLOT_REF slot)
-{
-    return slot.pool >= SAVEGAME_SLOT_POOL_NORMAL
-        && slot.pool < SAVEGAME_SLOT_POOL_NUMBER_OF && slot.index >= 0
-        && slot.index < Savegame_GetSlotCount(slot.pool);
-}
-
-int32_t Savegame_SlotToParam(const SAVEGAME_SLOT_REF slot)
-{
-    if (!Savegame_IsValidSlotRef(slot)) {
-        return -1;
-    }
-    const uint32_t packed = ((uint32_t)slot.pool << 31) | (uint32_t)slot.index;
-    return (int32_t)packed;
-}
-
-SAVEGAME_SLOT_REF Savegame_SlotFromParam(const int32_t param)
-{
-    if (param == -1) {
-        return Savegame_InvalidSlot();
-    }
-
-    const uint32_t packed = (uint32_t)param;
-    const SAVEGAME_SLOT_POOL pool = (packed >> 31) & 1;
-    const int32_t index = (int32_t)(packed & 0x7FFFFFFF);
-    return (SAVEGAME_SLOT_REF) {
-        .pool = pool,
-        .index = index,
-    };
-}
-
 void Savegame_Init(void)
 {
-    m_ResumeInfo = Memory_Alloc(
-        sizeof(RESUME_INFO)
-        * (GF_GetLevelTable(GFLT_MAIN)->count
-           + GF_GetLevelTable(GFLT_DEMOS)->count));
-
-    M_AllocSlots();
-
-    const GF_LEVEL_TABLE *const level_table = GF_GetLevelTable(GFLT_DEMOS);
-    for (int32_t i = 0; i < level_table->count; i++) {
-        RESUME_INFO *const resume_info =
-            Savegame_GetCurrentInfo(&level_table->levels[i]);
-        resume_info->lara_hitpoints = LARA_MAX_HITPOINTS;
-        resume_info->flags.available = true;
-        Inv_State_SetCount(&resume_info->inv, O_PISTOL_ITEM, 1);
-        resume_info->inv.ammo[LGT_PISTOLS] = Gun_GetInitialRounds(LGT_PISTOLS);
-        resume_info->gun_status = LGS_ARMLESS;
-        resume_info->equipped_gun_type = LGT_PISTOLS;
-        resume_info->holsters_gun_type = LGT_PISTOLS;
-        resume_info->back_gun_type = LGT_UNARMED;
-        resume_info->prev_level = -1;
-    }
-}
-
-bool Savegame_IsInitialised(void)
-{
-    return m_NormalSavegameInfo != nullptr;
+    Savegame_Resume_Init();
+    Savegame_Manager_Init();
 }
 
 void Savegame_Shutdown(void)
 {
-    M_FreeSlots();
-    Memory_FreePointer(&m_ResumeInfo);
-}
-
-void Savegame_ResizeSlots(void)
-{
-    M_FreeSlots();
-    M_AllocSlots();
-}
-
-int32_t Savegame_GetSlotCount(const SAVEGAME_SLOT_POOL pool)
-{
-    switch (pool) {
-    case SAVEGAME_SLOT_POOL_NORMAL:
-        return g_Config.gameplay.maximum_save_slots;
-    case SAVEGAME_SLOT_POOL_QUICK:
-        return g_Config.gameplay.maximum_quick_save_slots;
-    case SAVEGAME_SLOT_POOL_NUMBER_OF:
-        break;
-    }
-    return 0;
-}
-
-SAVEGAME_SLOT_REF Savegame_GetNextQuickSlot(void)
-{
-    if (m_QuickSaveSlots <= 0) {
-        return Savegame_InvalidSlot();
-    }
-    if (m_NextQuickSlot < 0 || m_NextQuickSlot >= m_QuickSaveSlots) {
-        m_NextQuickSlot = 0;
-    }
-    return Savegame_QuickSlot(m_NextQuickSlot);
-}
-
-int32_t Savegame_GetQuickVisualCount(void)
-{
-    int32_t count = 0;
-    const int32_t quick_slot_count =
-        Savegame_GetSlotCount(SAVEGAME_SLOT_POOL_QUICK);
-    for (int32_t i = 0; i < quick_slot_count; i++) {
-        if (!Savegame_IsSlotFree(Savegame_QuickSlot(i))) {
-            count++;
-        }
-    }
-    return count;
-}
-
-SAVEGAME_SLOT_REF Savegame_QuickFromVisualIndex(const int32_t visual_index)
-{
-    if (visual_index < 0) {
-        return Savegame_InvalidSlot();
-    }
-
-    const int32_t quick_slot_count =
-        Savegame_GetSlotCount(SAVEGAME_SLOT_POOL_QUICK);
-    for (int32_t i = 0; i < quick_slot_count; i++) {
-        const SAVEGAME_SLOT_REF candidate = Savegame_QuickSlot(i);
-        if (Savegame_IsSlotFree(candidate)) {
-            continue;
-        }
-
-        int32_t better_count = 0;
-        for (int32_t j = 0; j < quick_slot_count; j++) {
-            const SAVEGAME_SLOT_REF other = Savegame_QuickSlot(j);
-            if (Savegame_IsSlotFree(other)) {
-                continue;
-            }
-            if (M_IsQuickSlotSortedBefore(other, candidate)) {
-                better_count++;
-            }
-        }
-
-        if (better_count == visual_index) {
-            return candidate;
-        }
-    }
-
-    return Savegame_InvalidSlot();
-}
-
-int32_t Savegame_QuickToVisualIndex(const SAVEGAME_SLOT_REF slot)
-{
-    if (!Savegame_IsValidSlotRef(slot) || slot.pool != SAVEGAME_SLOT_POOL_QUICK
-        || Savegame_IsSlotFree(slot)) {
-        return -1;
-    }
-
-    const int32_t quick_slot_count =
-        Savegame_GetSlotCount(SAVEGAME_SLOT_POOL_QUICK);
-    int32_t better_count = 0;
-    for (int32_t i = 0; i < quick_slot_count; i++) {
-        const SAVEGAME_SLOT_REF other = Savegame_QuickSlot(i);
-        if (Savegame_IsSlotFree(other)) {
-            continue;
-        }
-        if (M_IsQuickSlotSortedBefore(other, slot)) {
-            better_count++;
-        }
-    }
-    return better_count;
-}
-
-RESUME_INFO *Savegame_GetCurrentInfo(const GF_LEVEL *const level)
-{
-    ASSERT(m_ResumeInfo != nullptr);
-    if (level == nullptr) {
-        return nullptr;
-    }
-    if (GF_GetLevelTableType(level->type) == GFLT_MAIN) {
-        return &m_ResumeInfo[level->num];
-    } else if (level->type == GFL_DEMO) {
-        return &m_ResumeInfo[GF_GetLevelTable(GFLT_MAIN)->count];
-    } else if (level->type == GFL_CUTSCENE || level->type == GFL_TITLE) {
-        return nullptr;
-    }
-    LOG_WARNING(
-        "Warning: unable to get resume info for level %d (type=%s)", level->num,
-        ENUM_MAP_TO_STRING(GF_LEVEL_TYPE, level->type));
-    return nullptr;
-}
-
-void Savegame_SetCurrentInfo(const int32_t current_slot, const int32_t src_slot)
-{
-    m_ResumeInfo[current_slot] = m_ResumeInfo[src_slot];
-}
-
-const SAVEGAME_INFO *Savegame_GetSavegameInfo(const SAVEGAME_SLOT_REF slot)
-{
-    return M_GetSavegameInfoSlot(slot);
-}
-
-void Savegame_InitCurrentInfo(void)
-{
-    const GF_LEVEL_TABLE *const level_table = GF_GetLevelTable(GFLT_MAIN);
-    for (int32_t i = 0; i < level_table->count; i++) {
-        const GF_LEVEL *const level = &level_table->levels[i];
-        Savegame_ResetCurrentInfo(level);
-        Savegame_ApplyLogicToCurrentInfo(level);
-        RESUME_INFO *const current = Savegame_GetCurrentInfo(level);
-        current->level_completed = false;
-        current->flags.available = false;
-    }
-
-    if (GF_GetGymLevel() != nullptr) {
-        Savegame_GetCurrentInfo(GF_GetGymLevel())->flags.available = true;
-    }
-    if (GF_GetFirstLevel() != nullptr) {
-        Savegame_GetCurrentInfo(GF_GetFirstLevel())->flags.available = true;
-    }
-
-    // Lara's wetness survives level transitions; a fresh playthrough starts
-    // dry.
-    LARA_INFO *const lara = Lara_GetLaraInfo();
-    memset(lara->wet, 0, sizeof(lara->wet));
-
-    // The rules last as long as the playthrough too; a load restores the saved
-    // ones over these once the file is read.
-    Rules_Reset();
-
-    // Which cutscenes have run lasts as long as the playthrough as well, so a
-    // fresh one sees all of them again.
-    CutSeq_SetPlayedMask(0);
-}
-
-void Savegame_ResetCurrentInfo(const GF_LEVEL *const level)
-{
-    LOG_INFO("Resetting resume info for level #%d", level->num);
-    RESUME_INFO *const current = Savegame_GetCurrentInfo(level);
-    *current = (RESUME_INFO) { .prev_level = -1, .level_completed = false };
-}
-
-void Savegame_CarryCurrentInfoToNextLevel(
-    const GF_LEVEL *const src_level, const GF_LEVEL *const dst_level)
-{
-    LOG_INFO(
-        "Copying resume info from level #%d to level #%d", src_level->num,
-        dst_level->num);
-    RESUME_INFO *const src_resume = Savegame_GetCurrentInfo(src_level);
-    RESUME_INFO *const dst_resume = Savegame_GetCurrentInfo(dst_level);
-    if (src_resume != nullptr && dst_resume != nullptr) {
-        const bool dst_level_completed = dst_resume->level_completed;
-        M_CopyResumeInfo(dst_resume, src_resume);
-        dst_resume->level_completed = dst_level_completed;
-        dst_resume->prev_level = src_level->num;
-    }
-}
-
-void Savegame_PersistGameToCurrentInfo(const GF_LEVEL *const level)
-{
-    RESUME_INFO *const resume = Savegame_GetCurrentInfo(level);
-    if (resume == nullptr) {
-        return;
-    }
-
-    resume->flags.available = true;
-
-    LARA_INFO *const lara = Lara_GetLaraInfo();
-    const ITEM *const lara_item = Lara_GetItem();
-
-    if (lara_item != nullptr) {
-        resume->lara_hitpoints = lara_item->hit_points;
-    }
-    resume->burning = g_TRVersion >= 4 && lara->burn;
-    M_PersistInventory(resume);
-
-    resume->equipped_gun_type = lara->last_gun_type;
-    resume->holsters_gun_type = lara->holsters_gun_type;
-    resume->back_gun_type = lara->back_gun_type;
-    if (resume->back_gun_type == LGT_UNARMED
-        && Gun_IsRifleType(resume->equipped_gun_type)
-        && Inv_HasItem(Gun_GetGunObject(resume->equipped_gun_type))) {
-        // If a rifle is currently drawn, Lara's back mesh is temporarily
-        // unarmed. Preserve the preferred rifle for next-level mesh restore.
-        resume->back_gun_type = resume->equipped_gun_type;
-    }
-    if (lara->gun_status == LGS_READY) {
-        resume->gun_status = LGS_READY;
-    } else {
-        resume->gun_status = LGS_ARMLESS;
-    }
-}
-
-void Savegame_ApplyLogicToCurrentInfo(const GF_LEVEL *const level)
-{
-    RESUME_INFO *const resume = Savegame_GetCurrentInfo(level);
-    if (resume == nullptr) {
-        return;
-    }
-
-    LOG_INFO("Applying game logic to level #%d", level->num);
-
-    if (!g_Config.gameplay.disable_healing_between_levels
-        || level == GF_GetGymLevel() || level == GF_GetFirstLevel()) {
-        resume->lara_hitpoints = g_Config.gameplay.start_lara_hitpoints;
-    }
-
-    if (level == GF_GetGymLevel()) {
-        resume->flags.available = true;
-        resume->flags.costume = g_TRVersion == 1;
-
-        // The gym is a house tour: she arrives with nothing.
-        resume->inv = (INVENTORY_STATE) {};
-
-        resume->equipped_gun_type = LGT_UNARMED;
-        resume->holsters_gun_type = LGT_UNARMED;
-        resume->back_gun_type = LGT_UNARMED;
-        resume->gun_status = LGS_ARMLESS;
-    }
-
-    if (level == GF_GetFirstLevel()) {
-        resume->flags.available = true;
-        resume->flags.costume = false;
-
-        // She starts the game with her pistols and nothing else.
-        resume->inv = (INVENTORY_STATE) {};
-        Inv_State_SetCount(&resume->inv, O_PISTOL_ITEM, 1);
-        resume->inv.ammo[LGT_PISTOLS] = Gun_GetInitialRounds(LGT_PISTOLS);
-
-        resume->equipped_gun_type = LGT_PISTOLS;
-        resume->holsters_gun_type = LGT_PISTOLS;
-        resume->back_gun_type = LGT_UNARMED;
-        resume->gun_status = LGS_ARMLESS;
-    }
-
-    if (Game_IsBonusFlagSet(GBF_NGPLUS) && level != GF_GetGymLevel()) {
-        // A bonus game hands her every weapon the game has, loaded.
-        for (LARA_GUN_TYPE gun_type = LGT_UNARMED + 1; gun_type < NUM_WEAPONS;
-             gun_type++) {
-            const OBJECT_ID gun_object = Gun_GetGunObject(gun_type);
-            if (gun_object == NO_OBJECT || !g_Weapons[gun_type].is_available) {
-                continue;
-            }
-            Inv_State_SetCount(&resume->inv, gun_object, 1);
-            resume->inv.ammo[gun_type] = gun_type == LGT_PISTOLS
-                ? Gun_GetInitialRounds(LGT_PISTOLS)
-                : 10000;
-        }
-        if (g_TRVersion > 1) {
-            Inv_State_SetCount(&resume->inv, O_FLARE_ITEM, MAX_QTY);
-        }
-
-        const bool should_force_ngplus_gun_setup =
-            !g_Config.gameplay.remember_gun_status || resume->prev_level == -1;
-        if (should_force_ngplus_gun_setup) {
-            switch (g_TRVersion) {
-            case 1:
-                resume->equipped_gun_type = LGT_UZIS;
-                resume->back_gun_type = LGT_SHOTGUN;
-                resume->holsters_gun_type = LGT_UZIS;
-                break;
-            case 2:
-                resume->equipped_gun_type = LGT_GRENADE;
-                resume->back_gun_type = LGT_GRENADE;
-                resume->holsters_gun_type = LGT_PISTOLS;
-                break;
-            case 3:
-            default:
-                resume->equipped_gun_type = LGT_ROCKET;
-                resume->back_gun_type = LGT_ROCKET;
-                resume->holsters_gun_type = LGT_PISTOLS;
-                break;
-            }
-        }
-    }
-
-    resume->stats.secret_flags = 0;
-
-    M_DetermineLegacyGunTypes(resume);
-}
-
-int32_t Savegame_GetCompletedLevelCount(void)
-{
-    int32_t count = 0;
-    const GF_LEVEL_TABLE *const level_table = GF_GetLevelTable(GFLT_MAIN);
-    for (int32_t i = 0; i < level_table->count; i++) {
-        const GF_LEVEL *const level = &level_table->levels[i];
-        const RESUME_INFO *const current = Savegame_GetCurrentInfo(level);
-        if (current != nullptr && current->level_completed) {
-            count++;
-        }
-    }
-    return count;
+    Savegame_Manager_Shutdown();
+    Savegame_Resume_Shutdown();
 }
 
 bool Savegame_IsManualSaveAllowed(void)
@@ -909,72 +100,14 @@ void Savegame_ProcessItemsBeforeLoad(void)
     }
 }
 
-void Savegame_ScanSavedGames(void)
-{
-    BENCHMARK benchmark = Benchmark_Start();
-    M_ClearSlots();
-
-    m_SaveCounter = 0;
-    m_SavedGames = 0;
-    m_MostRecentlyCreatedSlot = Savegame_InvalidSlot();
-    m_NextQuickSlot = 0;
-    int32_t newest_quick_counter = -1;
-    int32_t newest_quick_slot = -1;
-
-    // Scan low-priority locations first; the write directory is authoritative.
-    M_ScanSavedGamesDir(".");
-    M_ScanSavedGamesDir(TRXPath_Get(TRX_PATH_LEGACY_SAVES_DIR));
-    M_ScanSavedGamesDir(TRXPath_Get(TRX_PATH_SAVES_DIR));
-
-    {
-        // M_GetSaveWriteDir may use static formatting storage, so copy it
-        // before scanning because nested formatting calls during scan can
-        // overwrite it.
-        AUTO_FREE char *write_dir = Memory_DupStr(M_GetSaveWriteDir());
-        M_ScanSavedGamesDir(write_dir);
-    }
-
-    for (SAVEGAME_SLOT_POOL pool = 0; pool < SAVEGAME_SLOT_POOL_NUMBER_OF;
-         pool++) {
-        for (int32_t i = 0; i < Savegame_GetSlotCount(pool); i++) {
-            SAVEGAME_INFO *const savegame_info = M_GetSavegameInfoSlot(
-                (SAVEGAME_SLOT_REF) { .pool = pool, .index = i });
-            if (savegame_info->level_title == nullptr) {
-                continue;
-            }
-            if (savegame_info->counter > m_SaveCounter) {
-                m_SaveCounter = savegame_info->counter;
-                m_MostRecentlyCreatedSlot =
-                    (SAVEGAME_SLOT_REF) { .pool = pool, .index = i };
-            }
-            m_SavedGames++;
-
-            if (pool == SAVEGAME_SLOT_POOL_QUICK
-                && savegame_info->counter > newest_quick_counter) {
-                newest_quick_counter = savegame_info->counter;
-                newest_quick_slot = i;
-            }
-        }
-    }
-
-    if (m_QuickSaveSlots > 0 && newest_quick_slot >= 0) {
-        m_NextQuickSlot = (newest_quick_slot + 1) % m_QuickSaveSlots;
-    }
-
-    Benchmark_End(&benchmark, nullptr);
-}
-
 bool Savegame_Save(const SAVEGAME_SLOT_REF slot)
 {
     if (!Savegame_IsValidSlotRef(slot)) {
         return false;
     }
-
-    bool result = false;
     Savegame_BindSlot(slot);
 
     const GF_LEVEL *const current_level = Game_GetCurrentLevel();
-
     Savegame_PersistGameToCurrentInfo(current_level);
 
     const GF_LEVEL_TABLE *const level_table = GF_GetLevelTable(GFLT_MAIN);
@@ -985,76 +118,7 @@ bool Savegame_Save(const SAVEGAME_SLOT_REF slot)
         }
     }
 
-    SAVEGAME_INFO *const savegame_info = M_GetSavegameInfoSlot(slot);
-    const bool was_slot_empty = savegame_info->full_path == nullptr;
-
-    m_SaveCounter++;
-    const char *const save_pattern = M_GetSaveFilePatternForPool(slot.pool);
-    char *file_name = String_Format(save_pattern, slot.index);
-    char *full_path = M_GetSaveWritePath(file_name);
-    File_EnsureParentDirectories(full_path);
-    MYFILE *const fp = File_Open(full_path, FILE_OPEN_WRITE);
-    if (fp != nullptr) {
-        savegame_info->is_quick = slot.pool == SAVEGAME_SLOT_POOL_QUICK;
-        SG_File_SaveToFile(fp, savegame_info);
-        File_Close(fp);
-        result = true;
-    }
-    if (result) {
-        M_FillSlot(slot, full_path);
-    }
-
-    Memory_FreePointer(&file_name);
-    Memory_FreePointer(&full_path);
-
-    if (result) {
-        m_MostRecentlyCreatedSlot = slot;
-        if (was_slot_empty) {
-            m_SavedGames++;
-        }
-
-        if (slot.pool == SAVEGAME_SLOT_POOL_QUICK && m_QuickSaveSlots > 0) {
-            m_NextQuickSlot = (slot.index + 1) % m_QuickSaveSlots;
-        }
-    } else {
-        m_SaveCounter--;
-    }
-
-    return result;
-}
-
-bool Savegame_Delete(const SAVEGAME_SLOT_REF slot)
-{
-    if (!Savegame_IsValidSlotRef(slot) || Savegame_IsSlotFree(slot)) {
-        return false;
-    }
-
-    SAVEGAME_INFO *const savegame_info = M_GetSavegameInfoSlot(slot);
-    if (savegame_info == nullptr || savegame_info->full_path == nullptr) {
-        return false;
-    }
-
-    const bool result = remove(savegame_info->full_path) == 0;
-    if (!result) {
-        return false;
-    }
-
-    M_ClearSlot(savegame_info);
-    if (m_SavedGames > 0) {
-        m_SavedGames--;
-    }
-    if (m_BoundSlot.pool == slot.pool && m_BoundSlot.index == slot.index) {
-        m_BoundSlot = Savegame_InvalidSlot();
-    }
-    if (m_MostRecentlyUsedSlot.pool == slot.pool
-        && m_MostRecentlyUsedSlot.index == slot.index) {
-        m_MostRecentlyUsedSlot = Savegame_InvalidSlot();
-    }
-    if (m_MostRecentlyCreatedSlot.pool == slot.pool
-        && m_MostRecentlyCreatedSlot.index == slot.index) {
-        m_MostRecentlyCreatedSlot = Savegame_InvalidSlot();
-    }
-    return true;
+    return Savegame_Manager_WriteSlot(slot);
 }
 
 bool Savegame_Load(const SAVEGAME_SLOT_REF slot)
