+++ src/trx/game/savegame/vars.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <trx/game/savegame/types.h>
+
+// The names savegames give to what they carry. A key here is what appears in
+// the file, so an entry may be renamed only alongside the readers that accept
+// the old spelling.
+
+extern const SAVEGAME_INVENTORY_ENTRY g_Savegame_InventoryItems[];
+extern const SAVEGAME_AMMO_ENTRY g_Savegame_WeaponAmmo[];
+extern const SAVEGAME_RESUME_WEAPON g_Savegame_ResumeWeapons[];
+extern const SAVEGAME_RESUME_ITEM g_Savegame_ResumeItems[];
